pyproject.toml
[project]
name = "essay-scorer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.118.0",
    "google-genai>=1.41.0",
    "openai>=2.1.0",
    "pandas>=2.3.3",
    "requests>=2.32.5",
]
